+++ fictionops/src/fictionops/core.py
@@ -13,6 +13,7 @@ from .agent_revise_workflow import *
 from .agent_revision_accept import *
 from .agent_revision_runtime import *
 from .agent_project_context import *
+from .agent_preservation_verifier import *
 from .agent_policy import *
 from .agent_research_baseline import *
 from .agent_comprehensive_review import *